// core/src/component/component.types.d.ts
export type CustomActionModel = {
    type?: 'Custom' | undefined;
    package?: string;
    name: string;
    description?: string;
    data?: string | number | boolean | Formula;
    arguments?: {
        name: string;
        formula: Formula;
    }[];
    events?: Record<string, {
        actions: ActionModel[];
    }>;
    version?: 2 | never;
};
export type SwitchActionModel = {
    type: 'Switch';
    data?: string | number | boolean | Formula;
    cases: Array<{
        condition: Formula;
        actions: ActionModel[];
    }>;
    default: {
        actions: ActionModel[];
    };
};
export type VariableActionModel = {
    type: 'SetVariable';
    variable: string;
    data: Formula;
};
export type FetchActionModel = {
    type: 'Fetch';
    api: string;
    inputs?: Record<string, {
        formula: Formula;
    }>;
    onSuccess: {
        actions: ActionModel[];
    };
    onError: {
        actions: ActionModel[];
    };
    onMessage?: {
        actions: ActionModel[];
    };
};
export type SetURLParameterAction = {
    type: 'SetURLParameter';
    parameter: string;
    data: Formula;
};
export type EventActionModel = {
    type: 'TriggerEvent';
    event: string;
    data: Formula;
};
export type WorkflowActionModel = {
    type: 'TriggerWorkflow';
    workflow: string;
    parameters: Record<string, {
        formula: Formula;
    }>;
    contextProvider?: string;
};
export type ActionModel = VariableActionModel | EventActionModel | SwitchActionModel | FetchActionModel | CustomActionModel | SetURLParameterAction | WorkflowActionModel;